Script Cleanup

+++ Untitled RPG 3D/Assets/Scripts/Game Scripts/ScriptableObjects/Scripts/EnemyScriptableObject.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Enemy", menuName = "Enemy")]
 public class EnemyScriptableObject : ScriptableObject
 {
-    public int maxHP;
+    public float maxHP;
     public float speed;
     public float maxTime = 1.0f;
     public float maxDistance = 1.0f;
